Open an ODBC channel to a SQL Server database

Creates & opens a channel and checks its important characteristics.

open_dsn_channel_sqls(dsn_name,
  driver_version_minimum = numeric_version("13.0"),
  driver_version_maximum = numeric_version("99.0"))

Arguments

dsn_name
Name of the locally-defined DSN passed to RODBC::odbcConnect().

driver_version_minimum
The driver must be at least this version number. Represented as a base::numeric_version()

driver_version_maximum
The driver must not exceed this version number. Represented as a base::numeric_version()

Details

A DSN channel requires more code than usual to diagnose problems, because the DSN is defined on the local computer, and is not under the control of the repository. This function wraps the basic RODBC::odbcConnect() function with some checks. If unsuccessful, it returns some hints how to correct the problem, such as downloading the newest version from the Microsoft website.

Note

Assuring a minimum version is important, because driver versions can interpret values differently. For example, earlier version (before 11.0) returned dates as characters, which would propogate undetected through our code until it broke something with an unhelpful error message.

Examples

# NOT RUN {
requireNamespace("OuhscMunge")

OuhscMunge::open_dsn_channel_sqls(
  dsn_name        = "miechv_eval"
)
# }