Upload to a SQL Server database using RODBC

The function performs some extra configuration to improve robustness.

upload_sqls_rodbc(d, table_name, dsn_name, schema_name = NA_character_,
  clear_table = FALSE, create_table = FALSE,
  convert_logical_to_integer = FALSE, transaction = FALSE,
  verbose = TRUE)

Arguments

d
Dataset to be uploaded to the dataset. The object must inherit from data.frame.

table_name
Name of the database destination table. Can include the schema.

dsn_name
Name of the locally-defined DSN passed to RODBC::odbcConnect().

schema_name
Name of the database destination table. If it's not NA, the table_name will be qualified with it.

clear_table
If TRUE, calls RODBC::sqlClear() before writing to the table.

create_table
If the table structure has not yet been defined in the database, it will be created if create_table is TRUE.

convert_logical_to_integer
Convert all logical columns to integer. This helps the database store the values as bits.

transaction
Should the clear and upload steps be wrapped in a rollback transaction?

verbose
Write a message about the status of a successful upload.

Details

If transaction is TRUE and the upload fails, the table is rolled back to the state before function was callled. This includes rolling back the (optional) clearing of records, and uploading the new records. Decide if it's more robust to rollback to the previous state, or if it's better to leave the table in the incomplete state. The latter is helpful diagnosing which record caused the write to fail; look at the last successful record contained in the database

Examples

# NOT RUN {
requireNamespace("OuhscMunge")

OuhscMunge::upload_sqls_rodbc(
  d                          = ds_client,          # Some data.frame that exists in RAM
  table_name                 = "tbl_client",
  dsn_name                   = "miechv_eval",
  create_table               = FALSE,
  clear_table                = TRUE,
  transaction                = TRUE,
  verbose                    = TRUE,
  convert_logical_to_integer = TRUE
)
# }