Generates code to verify a data.frame.

Inspects properties of a data.frame and prints code to the console that a developer can use to start to check the properties of a dataset, such as the names and types of each column.

verify_value_headstart(d)

Arguments

d
The data.frame to verify. Required.

See also

Examples

library(magrittr)
verify_value_headstart(datasets::OrchardSprays)

#> checkmate::assert_numeric( ds$decrease  , any.missing=F , lower=2, upper=130 )
#> checkmate::assert_numeric( ds$rowpos    , any.missing=F , lower=1, upper=8   )
#> checkmate::assert_numeric( ds$colpos    , any.missing=F , lower=1, upper=8   )
#> checkmate::assert_factor(  ds$treatment , any.missing=F                      )

verify_value_headstart(datasets::iris)

#> checkmate::assert_numeric( ds$Sepal.Length , any.missing=F , lower=4, upper=8 )
#> checkmate::assert_numeric( ds$Sepal.Width  , any.missing=F , lower=2, upper=5 )
#> checkmate::assert_numeric( ds$Petal.Length , any.missing=F , lower=1, upper=7 )
#> checkmate::assert_numeric( ds$Petal.Width  , any.missing=F , lower=0, upper=3 )
#> checkmate::assert_factor(  ds$Species      , any.missing=F                    )

verify_value_headstart(datasets::BOD)

#> checkmate::assert_numeric( ds$Time   , any.missing=F , lower=1, upper=7  , unique=T)
#> checkmate::assert_numeric( ds$demand , any.missing=F , lower=8, upper=20 , unique=T)

verify_value_headstart(dplyr::band_members)

#> checkmate::assert_character(ds$name , any.missing=F , pattern="^.{4,4}$" , unique=T)
#> checkmate::assert_character(ds$band , any.missing=F , pattern="^.{6,7}$" )

storms_2 <- dplyr::storms %>%
  dplyr::mutate(
    storm_date = as.Date(ISOdate(year, month, day))
  )
verify_value_headstart(storms_2)

#> checkmate::assert_character(ds$name        , any.missing=F , pattern="^.{3,9}$"                                       )
#> checkmate::assert_numeric(  ds$year        , any.missing=F , lower=1975, upper=2015                                   )
#> checkmate::assert_numeric(  ds$month       , any.missing=F , lower=1, upper=12                                        )
#> checkmate::assert_integer(  ds$day         , any.missing=F , lower=1, upper=31                                        )
#> checkmate::assert_numeric(  ds$hour        , any.missing=F , lower=0, upper=23                                        )
#> checkmate::assert_numeric(  ds$lat         , any.missing=F , lower=7, upper=52                                        )
#> checkmate::assert_numeric(  ds$long        , any.missing=F , lower=-110, upper=-6                                     )
#> checkmate::assert_character(ds$status      , any.missing=F , pattern="^.{9,19}$"                                      )
#> checkmate::assert_ordered(  ds$category    , any.missing=F                                                            )
#> checkmate::assert_integer(  ds$wind        , any.missing=F , lower=10, upper=160                                      )
#> checkmate::assert_integer(  ds$pressure    , any.missing=F , lower=882, upper=1022                                    )
#> checkmate::assert_numeric(  ds$ts_diameter , any.missing=T , lower=0, upper=1002                                      )
#> checkmate::assert_numeric(  ds$hu_diameter , any.missing=T , lower=0, upper=346                                       )
#> checkmate::assert_date(     ds$storm_date  , any.missing=F , lower=as.Date("1975-06-27"), upper=as.Date("2015-11-11") )