Object inherits from data.frame. (Soft-deprecated.)

Check that the object inherits from data.frame. If not, throw an error.

This function has been soft-deprecated (i.e., it works, but with a warning). Please use the superior checkmate functions, checkmate::assert_data_frame() and checkmate::assert_tibble()

This helps check database-reading functions (e.g., RODBC::sqlQuery()) that return a data.frame if successful, and a character vector is unsucessful.

A minimum row count can be used to check that a trivially small number of rows was not returned. If minimum_row_count is set to 0, the function is similar to checkmate::assert_class(d, "data.frame"), but with with a more specific error message.

verify_data_frame(d, minimum_row_count = 10L)

Arguments

d
The object to verify. Required.

minimum_row_count
The data.frame should have at least this many rows. Defaults to 10. The datatype does not have to be an integer, but should be safely convertible to an integer.

See also

Examples

verify_data_frame(datasets::OrchardSprays, 20)

#> Warning: OuhscMunge::verify_data_frame() is deprecated.
#> Instead, please use checkmate::assert_data_frame()
#> with an appropriate value for the `min.rows` parameter.

verify_data_frame(datasets::iris, 4)

#> Warning: OuhscMunge::verify_data_frame() is deprecated.
#> Instead, please use checkmate::assert_data_frame()
#> with an appropriate value for the `min.rows` parameter.